feat: some structure
