+++ packages/api/package.json
@@ -10,6 +10,8 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@vendia/serverless-express": "4.5.4",
+    "aws-lambda": "1.0.7",
     "aws-sdk": "2.1110.0",
     "cors": "2.8.5",
     "express": "4.17.3",
