+++ packages/api/src/routes/post.routes.ts
@@ -1,0 +1,9 @@
+import express from 'express';
+
+const router = express.Router();
+
+router.get('/', async (req, res) => {
+  res.json({ 'post': 'route' });
+});
+
+export default router;
