+++ packages/api/src/routes/user.routes.ts
@@ -1,0 +1,9 @@
+import express from 'express';
+
+const router = express.Router();
+
+router.get('/', async (req, res) => {
+  res.json({ 'user': 'route' });
+});
+
+export default router;
