+++ packages/api/src/types/Post.ts
@@ -1,0 +1,3 @@
+export class Post {
+  id: string;
+}
