+++ packages/api/src/types/User.ts
@@ -1,0 +1,3 @@
+export class User {
+  id: string;
+}
